fix: 최상위 error boundary global-error로 수정
